backend/requirements-vast.txt
vastai==0.5.0
